# Conflicts:
#	.gitignore
#	sormas-api/pom.xml
#	sormas-api/src/main/java/de/symeda/sormas/api/contact/ContactDto.java
#	sormas-api/src/main/java/de/symeda/sormas/api/contact/ContactFacade.java
#	sormas-api/src/main/java/de/symeda/sormas/api/hospitalization/HospitalizationDto.java
#	sormas-api/src/main/java/de/symeda/sormas/api/i18n/Captions.java
#	sormas-api/src/main/java/de/symeda/sormas/api/i18n/Strings.java
#	sormas-api/src/main/java/de/symeda/sormas/api/sample/PathogenTestFacade.java
#	sormas-api/src/main/java/de/symeda/sormas/api/sample/SampleCriteria.java
#	sormas-api/src/main/java/de/symeda/sormas/api/sample/SampleFacade.java
#	sormas-api/src/main/java/de/symeda/sormas/api/sample/SampleIndexDto.java
#	sormas-api/src/main/resources/captions.properties
#	sormas-api/src/main/resources/strings.properties
#	sormas-app/app/src/main/java/de/symeda/sormas/app/backend/caze/Case.java
#	sormas-app/app/src/main/java/de/symeda/sormas/app/backend/common/DatabaseHelper.java
#	sormas-app/app/src/main/java/de/symeda/sormas/app/backend/contact/Contact.java
#	sormas-app/app/src/main/java/de/symeda/sormas/app/backend/hospitalization/Hospitalization.java
#	sormas-app/app/src/main/java/de/symeda/sormas/app/caze/edit/CaseEditFragment.java
#	sormas-app/app/src/main/java/de/symeda/sormas/app/caze/edit/CaseNewFragment.java
#	sormas-app/app/src/main/java/de/symeda/sormas/app/contact/edit/ContactEditFragment.java
#	sormas-app/app/src/main/res/layout/fragment_case_edit_layout.xml
#	sormas-app/app/src/main/res/layout/fragment_case_new_layout.xml
#	sormas-app/app/src/main/res/layout/fragment_contact_edit_layout.xml
#	sormas-backend/src/main/java/de/symeda/sormas/backend/caze/Case.java
#	sormas-backend/src/main/java/de/symeda/sormas/backend/common/BaseAdoService.java
#	sormas-backend/src/main/java/de/symeda/sormas/backend/contact/Contact.java
#	sormas-backend/src/main/java/de/symeda/sormas/backend/contact/ContactFacadeEjb.java
#	sormas-backend/src/main/java/de/symeda/sormas/backend/contact/ContactListCriteriaBuilder.java
#	sormas-backend/src/main/java/de/symeda/sormas/backend/hospitalization/Hospitalization.java
#	sormas-backend/src/main/java/de/symeda/sormas/backend/hospitalization/HospitalizationFacadeEjb.java
#	sormas-backend/src/main/java/de/symeda/sormas/backend/importexport/ImportFacadeEjb.java
#	sormas-backend/src/main/java/de/symeda/sormas/backend/sample/SampleFacadeEjb.java
#	sormas-backend/src/main/java/de/symeda/sormas/backend/sample/SampleService.java
#	sormas-keycloak-service-provider/pom.xml
#	sormas-ui/src/main/java/de/symeda/sormas/ui/UiUtil.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/caze/CaseController.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/caze/CaseCreateForm.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/caze/CaseDataForm.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/caze/components/linelisting/LineListingLayout.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/contact/ContactCreateForm.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/contact/ContactDataView.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/dashboard/DashboardDataProvider.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/dashboard/components/DashboardFilterLayout.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/hospitalization/HospitalizationForm.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/labmessage/LabMessageController.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/samples/AbstractSampleForm.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/samples/PathogenTestController.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/samples/PathogenTestForm.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/samples/SampleController.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/samples/SampleDataView.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/samples/SampleGridFilterForm.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/samples/pathogentestlink/PathogenTestList.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/samples/pathogentestlink/PathogenTestListComponent.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/utils/EpiWeekAndDateFilterComponent.java

+++ sormas-ui/src/main/java/de/symeda/sormas/ui/task/TaskController.java
@@ -28,26 +28,21 @@ import com.vaadin.ui.Notification.Type;
 import com.vaadin.ui.UI;
 import com.vaadin.ui.Window;
 
+import de.symeda.sormas.api.Disease;
 import de.symeda.sormas.api.FacadeProvider;
 import de.symeda.sormas.api.ReferenceDto;
-import de.symeda.sormas.api.caze.CaseDataDto;
-import de.symeda.sormas.api.contact.ContactDto;
-import de.symeda.sormas.api.event.EventDto;
 import de.symeda.sormas.api.i18n.I18nProperties;
 import de.symeda.sormas.api.i18n.Strings;
-import de.symeda.sormas.api.region.DistrictReferenceDto;
+import de.symeda.sormas.api.infrastructure.district.DistrictReferenceDto;
 import de.symeda.sormas.api.sample.SampleDto;
 import de.symeda.sormas.api.task.TaskContext;
 import de.symeda.sormas.api.task.TaskDto;
 import de.symeda.sormas.api.task.TaskIndexDto;
 import de.symeda.sormas.api.task.TaskType;
 import de.symeda.sormas.api.user.UserRight;
-import de.symeda.sormas.api.user.UserRole;
 import de.symeda.sormas.ui.ControllerProvider;
 import de.symeda.sormas.ui.UserProvider;
 import de.symeda.sormas.ui.utils.CommitDiscardWrapperComponent;
-import de.symeda.sormas.ui.utils.CommitDiscardWrapperComponent.CommitListener;
-import de.symeda.sormas.ui.utils.CommitDiscardWrapperComponent.DeleteListener;
 import de.symeda.sormas.ui.utils.VaadinUiUtil;
 
 public class TaskController {
@@ -56,33 +51,29 @@ public class TaskController {
 
 	}
 
-	public void create(TaskContext context, ReferenceDto entityRef, Runnable callback) {
+	public void create(TaskContext context, ReferenceDto entityRef, Disease disease, Runnable callback) {
 
-		TaskEditForm createForm = new TaskEditForm(true, false);
+		TaskEditForm createForm = new TaskEditForm(true, false, disease);
 		createForm.setValue(createNewTask(context, entityRef));
 		final CommitDiscardWrapperComponent<TaskEditForm> editView = new CommitDiscardWrapperComponent<TaskEditForm>(
 			createForm,
 			UserProvider.getCurrent().hasUserRight(UserRight.TASK_CREATE),
 			createForm.getFieldGroup());
 
-		editView.addCommitListener(new CommitListener() {
-
-			@Override
-			public void onCommit() {
-				if (!createForm.getFieldGroup().isModified()) {
-					TaskDto dto = createForm.getValue();
-					FacadeProvider.getTaskFacade().saveTask(dto);
-					callback.run();
-				}
+		editView.addCommitListener(() -> {
+			if (!createForm.getFieldGroup().isModified()) {
+				TaskDto dto = createForm.getValue();
+				FacadeProvider.getTaskFacade().saveTask(dto);
+				callback.run();
 			}
 		});
 
 		VaadinUiUtil.showModalPopupWindow(editView, I18nProperties.getString(Strings.headingCreateNewTask));
 	}
 
-	public void createSampleCollectionTask(TaskContext context, ReferenceDto entityRef, SampleDto sample) {
+	public void createSampleCollectionTask(TaskContext context, ReferenceDto entityRef, SampleDto sample, Disease disease) {
 
-		TaskEditForm createForm = new TaskEditForm(true, false);
+		TaskEditForm createForm = new TaskEditForm(true, false, disease);
 		TaskDto taskDto = createNewTask(context, entityRef);
 		taskDto.setTaskType(TaskType.SAMPLE_COLLECTION);
 		taskDto.setCreatorComment(sample.getNoTestPossibleReason());
@@ -93,61 +84,49 @@ public class TaskController {
 			createForm,
 			UserProvider.getCurrent().hasUserRight(UserRight.TASK_CREATE),
 			createForm.getFieldGroup());
-		createView.addCommitListener(new CommitListener() {
-
-			@Override
-			public void onCommit() {
-				if (!createForm.getFieldGroup().isModified()) {
-					TaskDto dto = createForm.getValue();
-					FacadeProvider.getTaskFacade().saveTask(dto);
-				}
+		createView.addCommitListener(() -> {
+			if (!createForm.getFieldGroup().isModified()) {
+				TaskDto dto = createForm.getValue();
+				FacadeProvider.getTaskFacade().saveTask(dto);
 			}
 		});
 
 		VaadinUiUtil.showModalPopupWindow(createView, I18nProperties.getString(Strings.headingCreateNewTask));
 	}
 
-	public void edit(TaskIndexDto dto, Runnable callback, boolean editedFromTaskGrid) {
+	public void edit(TaskIndexDto dto, Runnable callback, boolean editedFromTaskGrid, Disease disease) {
 
 		// get fresh data
 		TaskDto newDto = FacadeProvider.getTaskFacade().getByUuid(dto.getUuid());
 
-		TaskEditForm form = new TaskEditForm(false, editedFromTaskGrid);
+		TaskEditForm form = new TaskEditForm(false, editedFromTaskGrid, disease);
 		form.setValue(newDto);
 		final CommitDiscardWrapperComponent<TaskEditForm> editView =
 			new CommitDiscardWrapperComponent<TaskEditForm>(form, UserProvider.getCurrent().hasUserRight(UserRight.TASK_EDIT), form.getFieldGroup());
 
 		Window popupWindow = VaadinUiUtil.showModalPopupWindow(editView, I18nProperties.getString(Strings.headingEditTask));
 
-		editView.addCommitListener(new CommitListener() {
+		editView.addCommitListener(() -> {
+			if (!form.getFieldGroup().isModified()) {
+				TaskDto dto1 = form.getValue();
+				FacadeProvider.getTaskFacade().saveTask(dto1);
 
-			@Override
-			public void onCommit() {
-				if (!form.getFieldGroup().isModified()) {
-					TaskDto dto = form.getValue();
-					FacadeProvider.getTaskFacade().saveTask(dto);
-
-					if (!editedFromTaskGrid && dto.getCaze() != null) {
-						ControllerProvider.getCaseController().navigateToCase(dto.getCaze().getUuid());
-					}
-
-					popupWindow.close();
-					callback.run();
+				if (!editedFromTaskGrid && dto1.getCaze() != null) {
+					ControllerProvider.getCaseController().navigateToCase(dto1.getCaze().getUuid());
 				}
+
+				popupWindow.close();
+				callback.run();
 			}
 		});
 
-		editView.addDiscardListener(() -> popupWindow.close());
+		editView.addDiscardListener(popupWindow::close);
 
-		if (UserProvider.getCurrent().hasUserRole(UserRole.ADMIN)) {
-			editView.addDeleteListener(new DeleteListener() {
-
-				@Override
-				public void onDelete() {
-					FacadeProvider.getTaskFacade().deleteTask(newDto);
-					UI.getCurrent().removeWindow(popupWindow);
-					callback.run();
-				}
+		if (UserProvider.getCurrent().hasUserRight(UserRight.TASK_DELETE)) {
+			editView.addDeleteListener(() -> {
+				FacadeProvider.getTaskFacade().deleteTask(newDto);
+				UI.getCurrent().removeWindow(popupWindow);
+				callback.run();
 			}, I18nProperties.getString(Strings.entityTask));
 		}
 	}
@@ -167,43 +146,19 @@ public class TaskController {
 				Type.WARNING_MESSAGE,
 				false).show(Page.getCurrent());
 		} else {
-			VaadinUiUtil.showDeleteConfirmationWindow(
-				String.format(I18nProperties.getString(Strings.confirmationDeleteTasks), selectedRows.size()),
-				new Runnable() {
-
-					public void run() {
-						for (TaskIndexDto selectedRow : selectedRows) {
-							FacadeProvider.getTaskFacade().deleteTask(FacadeProvider.getTaskFacade().getByUuid(selectedRow.getUuid()));
-						}
-						callback.run();
-						new Notification(
-							I18nProperties.getString(Strings.headingTasksDeleted),
-							I18nProperties.getString(Strings.messageTasksDeleted),
-							Type.HUMANIZED_MESSAGE,
-							false).show(Page.getCurrent());
+			VaadinUiUtil
+				.showDeleteConfirmationWindow(String.format(I18nProperties.getString(Strings.confirmationDeleteTasks), selectedRows.size()), () -> {
+					for (TaskIndexDto selectedRow : selectedRows) {
+						FacadeProvider.getTaskFacade().deleteTask(FacadeProvider.getTaskFacade().getByUuid(selectedRow.getUuid()));
 					}
+					callback.run();
+					new Notification(
+						I18nProperties.getString(Strings.headingTasksDeleted),
+						I18nProperties.getString(Strings.messageTasksDeleted),
+						Type.HUMANIZED_MESSAGE,
+						false).show(Page.getCurrent());
 				});
 		}
-	}
-
-	private DistrictReferenceDto getTaskIndexDistrict(TaskIndexDto taskDto) {
-		DistrictReferenceDto district = null;
-		if (taskDto.getCaze() != null) {
-			CaseDataDto caseDto = FacadeProvider.getCaseFacade().getCaseDataByUuid(taskDto.getCaze().getUuid());
-			district = caseDto.getDistrict();
-		} else if (taskDto.getContact() != null) {
-			ContactDto contactDto = FacadeProvider.getContactFacade().getContactByUuid(taskDto.getContact().getUuid());
-			if (contactDto.getRegion() != null && contactDto.getDistrict() != null) {
-				district = contactDto.getDistrict();
-			} else {
-				CaseDataDto caseDto = FacadeProvider.getCaseFacade().getCaseDataByUuid(contactDto.getCaze().getUuid());
-				district = caseDto.getDistrict();
-			}
-		} else if (taskDto.getEvent() != null) {
-			EventDto eventDto = FacadeProvider.getEventFacade().getEventByUuid(taskDto.getEvent().getUuid());
-			district = eventDto.getEventLocation().getDistrict();
-		}
-		return district;
 	}
 
 	public void showBulkTaskDataEditComponent(Collection<? extends TaskIndexDto> selectedTasks, Runnable callback) {
